+++ src/components/setting/developer/enums.ts
@@ -1,0 +1,18 @@
+export enum Route {
+  HOME = 'Settings',
+  PLUGINS = 'Plugins',
+}
+
+export enum Icons {
+  setlog = 'console',
+  update = 'update',
+  showlog = 'bug',
+  cache = 'floppy',
+  clearcache = 'delete-sweep',
+  clearOrphanCache = 'delete-empty',
+  crossfade = 'shuffle-variant',
+  fade = 'cosine-wave',
+  plugins = 'puzzle',
+  Tanaka = 'emoticon-devil',
+  ArtworkRes = 'quality-high',
+}
